fixed a NullPointerException in ReadMessageWindow.
set it to open all the folders in the store when the store is connected if
Pooka.openFoldersOnConnect is set to true.

also fixed a bug where multiple FolderInfos were being created for Folders.

and fixed the bug where messagesRemoved was named messageRemoved, thus
preventing updates when messages are deleted.